ports/esp32/main.c: Limit uPy heap size to a ratio of IDF heap size.
Before this change the largest free heap block was allocated for uPy heap.
ESP32 S2 and S3 has different memory layout and this method results
in so small remaining IDF heap that WiFi and other tasks breaks.

The heuristic ratio was chosen as an usable compromise betwen
uPy and IDF heap sizes for devices using WiFi and TLS
and fine tuned to preserve original uPy heap size on ESP32 and IDF 4.0.

Allocation on SPIRAM-equipped devices is not changed.

Signed-off-by: Tomas Vanek <vanekt@fbl.cz>